[update] /new /edit毎にコンポーネントを用意
