+++ src/view/MemoList.vue
@@ -5,7 +5,7 @@
       <div>
         <tr>
           <td>
-            <router-link to="/new" @click="newMemoTextArea">＋ 新規メモ</router-link>
+            <router-link to="/new">＋ 新規メモ</router-link>
           </td>
         </tr>
       </div>
@@ -46,9 +46,6 @@
       },
       fetchFirstLine: function (contents) {
         return contents.split('\n')[0]
-      },
-      newMemoTextArea: function () {
-        store.commit('updateNewMemoTextArea', {boolean: true})
       }
     }
   }
